requirements.txt
pyinstaller==6.1.0
pyautogui==0.9.54
schedule==1.2.0
pynput==1.7.6
pystray==0.19.4
pillow==10.1.0
pywin32==306
opencv-python-headless==4.8.1.78